Revert "move logging back into shed-install because of weird double log issue"
This reverts commit ab6723fa36f6a55f01c131d1c61968fcfab0cbcf.
